RAG/requirements.txt
flask==2.3.3
openai==1.3.7
chromadb==0.4.18
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-dotenv==1.0.0
werkzeug==2.3.7
